return the enum instead of using send_data
for some rails apps using send_data resulted in `.to_s` being called on
the stream and that being send to the user (`"<CSVEnumerator#1234>"`
instead of the actual content).

With just returning the stream in our rails renderers, rails will set
that as the response.body. We still need to find out how to handle extra
options (like headers). Also, it would be interesting to figure out what
the difference of the affected application was to create a test-case.